move Immunizations services from common to clinicalServices. Move Immunizations snapshot tests to ClinicalInfo.test.tsx. Update imports
