+++ containers/ecr-viewer/src/app/view-data/services/ecrSummaryService.tsx
@@ -34,9 +34,11 @@ import { DisplayDataProps } from "@/app/view-data/components/DataDisplay";
 import { ConditionSummary } from "@/app/view-data/components/EcrSummary";
 import { LabAccordion } from "@/app/view-data/components/LabAccordion";
 import {
-  returnImmunizations,
   returnProblemsTable,
 } from "@/app/view-data/components/common";
+import {
+  returnImmunizations,
+} from "@/app/view-data/services/clinicalService";
 
 import {
   evaluatePatientName,
